pyproject.toml
[tool.poetry]
name = "llm-app"
version = "0.1.0"
description = ""
authors = ["Jacob Cohen <79770350+jacob187@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
langchain = "^0.3.1"
streamlit = "^1.38.0"
python-dotenv = "^1.0.1"
langchain-anthropic = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

pythonpath = ["."]